backend/requirements.txt
Flask==2.0.1
google-generativeai==0.1.0
firebase-admin==5.0.0
Pillow==8.3.2
gunicorn==20.1.0
Flask-Cors==3.0.10

